Change plural to singular.
